+++ src/nvlib/model/odt/odt_r_outline.py
@@ -153,8 +153,10 @@ class OdtROutline(OdtReader):
                     if attr[1]:
                         self.novel.languageCode = attr[1]
                 if attr[0] == 'country':
-                    if attr[1]:
+                    if attr[1] and attr[1] != 'none':
                         self.novel.countryCode = attr[1]
+                    else:
+                        self.novel.countryCode = None
             return
 
         if tag == 's':